+++ CHANGELOG.md
@@ -32,6 +32,16 @@ Security - in case of vulnerabilities
 - Updated module dependencies to latest versions
 - Internal tracking ticket: CASMHMS-6396
 
+## [1.28.0] - 2025-04-10
+
+### Fixed
+
+Fixed various concurrency issues:
+
+- Added mutex to protect the KnownDevices map from concurrent writes
+- Extended use of the Redis active pipeline mutex into the gcloud, JAWS,
+  and simulator backend helpers
+
 ## [1.27.0] - 2025-04-04
 
 ### Added